+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 Features:
 ---------
 
+  * 'dbstack' builtin get current instruction calling tree.
+
   * [#16](http://github.com/Nelson-numerical-software/nelson/issues/16): lastwarn builtin (Last recorded warning message).
 
   * [#15](http://github.com/Nelson-numerical-software/nelson/issues/15): warning builtin was extended (state and identifier added).